Modified usart tx to use state machine and added common makefile
It prints the dump sort of correctly, but it prints double characters
sometimes.  This seems to be because the start_transmitting bit is
always set, but trying to change that makes it send characters without
waiting for them to be transmitted.  I don't yet know the right
pattern to be using here to make the interconnected modules work