Modified to use an async fifo between snooper and dumper
I also had to replace the generator in the memory module because it
wasn't synthenizing properly, and I don't know why

The interface to the ram wasn't working, which is why I replaced it
with a fifo.  The timing was off, but I know the fifo interface to
ram works.  Using the valid/ready signals for the dumper was great
but the snooper, it isn't the best because the snooper can't wait
for the output to be read.  It has to keep up with the bus speed.